Move DiffSummary out
